+++ Dockerfile
@@ -24,4 +24,4 @@ COPY --from=composer /usr/local/src/vendor /usr/local/vendor
 
 HEALTHCHECK --interval=30s --timeout=15s --start-period=60s --retries=3 CMD curl -s -H "Authorization: Bearer ${OPR_EXECUTOR_SECRET}" --fail http://127.0.0.1:80/v1/health
 
-CMD [ "php", "app/run.php" ]
+CMD [ "php", "app/http.php" ]
